api/requirements.txt
fastapi==0.112.2
langchain==0.2.14
chromadb==0.5.3
langchain_chroma==0.1.3
langchain_community==0.2.12
langchain_openai==0.1.22
unstructured[md]~=0.15.7
pydantic==2.8.2
python-dotenv==1.0.1
uvicorn==0.30.6
python-jose[cryptography]==3.3.0
